LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-pseudo-elements-gcs-cq-expected.txt


FAIL ::scroll-marker-group queries root originating element for size queries assert_equals: expected "pass" but got ""
FAIL ::scroll-button skips queries root originating element for size queries assert_equals: expected "pass" but got ""
FAIL ::scroll-marker-group skips originating element for size queries assert_equals: expected "pass" but got ""
FAIL ::scroll-button skips originating element for size queries assert_equals: expected "pass" but got ""
